Make addUser(s)ToDiscordRole async
Note that this explicitly does not parallelize adding multiple users to
roles. Discord's API is RESTful enough that it requires one request per
user, and sending them in parallel seems likely to fall afoul of rate
limits.

However, I did use this opportunity to make the function operate on
multiple user IDs in sequence instead of just one, since there's a lot
of shared state per invocation.
